// projects/shared-library/src/lib/shared/gantt/types.ts
export type GanttTaskValue<T> =
  | GanttTaskValueWithAllocation<T>
  | GanttTaskValueWithSubAllocation<T>;

export interface BaseTaskValue<T> {
  id: string | number;
  start_date: Date;
  end_date: Date;
  name: string;
  subtitle?: string;
  render?: 'split';
  parent?: string | number;
  type: string | number;
  $open?: boolean;
  hasChildren: boolean;
  isParent: boolean;
  overallocated?: boolean;
  bar_height?: number;
  row_height?: number;
  payload: T;
  classes?: string[];
}

export function withSuballocations<T>(
  task: GanttTaskValueWithSubAllocation<T> | GanttTaskValueWithAllocation<T>
): task is GanttTaskValueWithSubAllocation<T> {
  return (
    (task as GanttTaskValueWithSubAllocation<T>).subAllocations !== undefined
  );
}

export interface GanttTaskValueWithAllocation<T> extends BaseTaskValue<T> {
  allocation: number;
}

export interface GanttTaskValueWithSubAllocation<T> extends BaseTaskValue<T> {
  subAllocations: SubAllocation[];
}

export type GanttScaleService = {
  scale: Timelines;
  config(options?: GanttScaleOptions): {
    unit: string;
    step: number;
    format: (date: Date) => string;
    css?: (date: Date) => string;
  }[];
};

// will be required for custom scale
export type GanttScaleOptions = {};

export type ContextItemClickEvent<T> = {
  event: NbMenuItem;
  task: GanttTaskValue<T>;
};

export type GanttEvent<T> = {
  event: string;
  task: GanttTaskValue<T>;
};

export type GanttRenderOptions<T = AnyObject> = {
  contextItems: NbMenuItem[];
  contextTemplate?: TemplateRef<AnyObject>;
  viewContainerRef?: ViewContainerRef;
  columnName?: string;
  showKebab: boolean;
  showParentInitials: boolean;
  showChildInitials: boolean;
  columnComponent: Type<IColumnComponent<T>>;
  barComponent: Type<IBarComponent<T>>;
  columnWidth: number;
  resizer: boolean;
  searchPlaceholder?: string;
  showSearch: boolean;
  moveToToday: boolean;
  highlightRange?: [Date, Date];
  showOverallocatedIcon: boolean;
  contextItemFilter?: ContextItemFilter<T>;
  defaultScale: Timelines;
  markToday: boolean;
  showTooltip?: boolean;
  childIndent: boolean;
};

export type GanttAllocationFields = {
  startDate: Date;
  endDate: Date;
  allocation: number;
};

export type GanttLib = typeof gantt;

export enum Timelines {
  Weekly,
  Monthly,
  Quarterly,
  Custom,
}
// ...
export type IColumnComponent<T> = {
  item: GanttTaskValue<T>;
  contextItems: NbMenuItem[];
  active: boolean;
  showKebab: boolean;
  showParentInitials: boolean;
  showChildInitials: boolean;
  showOverallocatedIcon: boolean;
  contextItemFilter?: ContextItemFilter<T>;
};

export type IBarComponent<T> = {
  item: GanttTaskValue<T>;
};

export type SubAllocation<T = AnyObject> = {
  percent: number;
  allocation: number;
  classes?: string[];
} & T;

export type ContextItemFilter<T> = (item: GanttTaskValue<T>) => NbMenuItem[];